+++ src/engine/components/numericLineGenerator/index.js
@@ -7,6 +7,7 @@
 import { Container } from 'pixi.js';
 
 import { type NumberActor } from 'engine/components/numberGenerator';
+import { HALF } from 'constants/numbers';
 
 import arrowGenerator from './arrowGenerator';
 import lineGenerator, { type Line } from './lineGenerator';
@@ -20,15 +21,18 @@ export type NumericLine = {|
   receiveNumberAtPosition(number: NumberActor, position: number): void,
 |};
 const numericLineGenerator = (numbers: Array<number | null>, size: number): NumericLine => {
-  const leftArrow = arrowGenerator();
-  const rightArrow = arrowGenerator(true);
-  const line = lineGenerator(numbers, size);
   const view = new Container();
+  const leftArrow = arrowGenerator(size);
+  const rightArrow = arrowGenerator(size, true);
+  const line = lineGenerator(numbers, size);
 
   line.view.x = leftArrow.width;
   rightArrow.x = line.view.width + leftArrow.width;
 
   view.addChild(leftArrow, line.view, rightArrow);
+
+  // Pivot at center top
+  view.pivot.x = view.width / HALF;
 
   return {
     view,