+++ src/engine/components/numericLineGenerator/index.stories.jsx
@@ -5,6 +5,7 @@
  * @flow
  */
 import React from 'react';
+import wait from 'speculation/wait';
 
 import { storiesOf } from 'test/storybook-facades';
 import PixiWrapper from 'test/PixiWrapper';
@@ -21,8 +22,6 @@ const WIDTH_NUMERIC_LINE = 960;
 const HEIGHT_NUMERIC_LINE = 300;
 const WIDTH_ARROW = 104;
 const HEIGHT_ARROW = 204;
-const WIDTH_LINE = 640;
-const HEIGHT_LINE = 200;
 const QUARTER = 4;
 const NINE = 9;
 
@@ -31,7 +30,7 @@ storiesOf('engine.components.numericLine', module)
   // Arrows
   //////////////////////////////////
   .add('simple left arrow', () => {
-    const arrow = arrowGenerator();
+    const arrow = arrowGenerator(SIZE);
 
     arrow.x = WIDTH_ARROW / QUARTER;
     arrow.y = HEIGHT_ARROW / QUARTER;
@@ -39,7 +38,7 @@ storiesOf('engine.components.numericLine', module)
     return <PixiWrapper component={arrow} height={HEIGHT_ARROW} width={WIDTH_ARROW} />;
   })
   .add('simple right arrow', () => {
-    const arrow = arrowGenerator(true);
+    const arrow = arrowGenerator(SIZE, true);
 
     arrow.x = WIDTH_ARROW / QUARTER;
     arrow.y = HEIGHT_ARROW / QUARTER;
@@ -53,19 +52,21 @@ storiesOf('engine.components.numericLine', module)
     const line = lineGenerator([null, null, null, null, null, null, null, null], SIZE);
 
     line.view.x = 32;
-    line.view.y = HEIGHT_LINE / QUARTER;
+    line.view.y = 28;
 
-    return <PixiWrapper component={line.view} height={HEIGHT_LINE} width={WIDTH_LINE} />;
+    return (
+      <PixiWrapper component={line.view} height={line.view.height + SIZE} width={line.view.width + SIZE} />
+    );
   })
   .add('line with static numbers', () => {
     // eslint-disable-next-line no-magic-numbers
     const line = lineGenerator([0, 1, 2, 3, 4, 5, 6, 7], SIZE);
 
     line.view.x = 32;
-    line.view.y = HEIGHT_LINE / QUARTER;
+    line.view.y = 28;
 
     return (
-      <PixiWrapper component={line.view} height={HEIGHT_LINE} width={WIDTH_LINE} />
+      <PixiWrapper component={line.view} height={line.view.height + SIZE} width={line.view.width + SIZE} />
     );
   })
   .add('line with an animated number', () => {
@@ -73,12 +74,15 @@ storiesOf('engine.components.numericLine', module)
     const ten = numberGenerator(TEN, '0,0', SIZE);
 
     line.view.x = 32;
-    line.view.y = HEIGHT_LINE / QUARTER;
+    line.view.y = 28;
 
-    line.receiveNumberAtPosition(ten, ZERO);
+    (async () => {
+      await wait(500); // eslint-disable-line no-magic-numbers
+      await line.receiveNumberAtPosition(ten, ZERO);
+    })();
 
     return (
-      <PixiWrapper component={line.view} height={HEIGHT_LINE} width={WIDTH_LINE} />
+      <PixiWrapper component={line.view} height={line.view.height + SIZE + SIZE} width={line.view.width + SIZE} />
     );
   })
   //////////////////////////////////
@@ -90,7 +94,7 @@ storiesOf('engine.components.numericLine', module)
       SIZE,
     );
 
-    numericLine.view.x = (WIDTH_NUMERIC_LINE - numericLine.view.width) / HALF;
+    numericLine.view.x = WIDTH_NUMERIC_LINE / HALF;
     numericLine.view.y = (HEIGHT_NUMERIC_LINE - numericLine.view.height) / HALF;
 
     return (
@@ -102,9 +106,12 @@ storiesOf('engine.components.numericLine', module)
     const numericLine = numericLineGenerator([1, 2, 3, 4, 5, 6, 7, 8, 9, null], SIZE);
     const ten = numberGenerator(TEN, '0,0', SIZE);
 
-    numericLine.receiveNumberAtPosition(ten, NINE);
+    (async () => {
+      await wait(500); // eslint-disable-line no-magic-numbers
+      await numericLine.receiveNumberAtPosition(ten, NINE);
+    })();
 
-    numericLine.view.x = (WIDTH_NUMERIC_LINE - numericLine.view.width) / HALF;
+    numericLine.view.x = WIDTH_NUMERIC_LINE / HALF;
     numericLine.view.y = (HEIGHT_NUMERIC_LINE - numericLine.view.height) / HALF;
 
     return (